test: reorganize tests into submodule
